+++ apps/connect/src/App.tsx
@@ -8,10 +8,8 @@ import NavBar from "./components/atoms/NavBar";
 import NewsBar from "./components/atoms/NewsBar";
 import messageConfig from "./configs/messages";
 import { useQueryParams } from "./hooks/useQueryParams";
-import WormholeConnect, {
-  MAINNET,
-  TESTNET,
-} from "@wormhole-foundation/wormhole-connect";
+import { useFormatAssetParam } from "./hooks/useFormatAssetParam";
+import WormholeConnect from "@wormhole-foundation/wormhole-connect";
 import { eventHandler } from "./providers/telemetry";
 
 const defaultConfig: WormholeConnectConfig = {
@@ -21,25 +19,11 @@ const defaultConfig: WormholeConnectConfig = {
     eventHandler: eventHandler,
   }),
 };
-const tokensList =
-  defaultConfig.env === "mainnet" ? MAINNET.tokens : TESTNET.tokens;
+
 export default function Root() {
   const { txHash, sourceChain, targetChain, asset, requiredNetwork } =
     useQueryParams();
-
-  let tokenKey: string | undefined = undefined;
-  const allTokens = {
-    ...tokensList,
-    ...defaultConfig?.tokensConfig,
-  };
-  if (allTokens) {
-    const tokenParam = Object.values(allTokens).find(
-      (config) => config?.tokenId?.address === asset || config?.key === asset
-    );
-    if (tokenParam) {
-      tokenKey = tokenParam.key;
-    }
-  }
+  const tokenKey = useFormatAssetParam(asset);
   const config = useMemo(
     () => ({
       ...defaultConfig,
