+++ apps/connect/src/hooks/useFormatAssetParam.ts
@@ -1,0 +1,28 @@
+import { MAINNET, TESTNET } from "@wormhole-foundation/wormhole-connect";
+import { useMemo } from "react";
+
+const tokensList =
+  wormholeConnectConfig.env === "mainnet" ? MAINNET.tokens : TESTNET.tokens;
+function getFormatedAsset(asset: string | null): string | null {
+  const allTokens = {
+    ...tokensList,
+    ...wormholeConnectConfig?.tokensConfig,
+  };
+  if (allTokens && asset) {
+    const tokenParam = Object.values(allTokens).find(
+      (config) => config?.tokenId?.address === asset || config?.key === asset
+    );
+    if (tokenParam) {
+      return tokenParam.key;
+    }
+  }
+  return null;
+}
+
+export function useFormatAssetParam(asset: string | null) {
+  const formatedAsset = useMemo(
+    () => getFormatedAsset(asset),
+    [asset]
+  );
+  return formatedAsset;
+}
